add async mode and refactor hedera utils
