+++ cli.js
@@ -24,10 +24,11 @@ yargs(hideBin(process.argv))
       CliOptions.addDevModeOption(yargs);
       CliOptions.addFullModeOption(yargs);
       CliOptions.addBalanceOption(yargs);
+      CliOptions.addAsyncOption(yargs);
     },
     async (argv) => {
       await NodeController.startLocalNode(argv.network, argv.limits, argv.dev, argv.full);
-      await main(argv.accounts, argv.balance, argv.detached, argv.host);
+      await main(argv.accounts, argv.async, argv.balance, argv.detached, argv.host);
     }
   )
   .command(
@@ -49,11 +50,12 @@ yargs(hideBin(process.argv))
       CliOptions.addDevModeOption(yargs);
       CliOptions.addFullModeOption(yargs);
       CliOptions.addBalanceOption(yargs);
+      CliOptions.addAsyncOption(yargs);
     },
     async (argv) => {
       await NodeController.stopLocalNode();
       await NodeController.startLocalNode(argv.network, argv.limits, argv.dev, argv.full);
-      await main(argv.accounts, argv.balance, argv.detached, argv.host);
+      await main(argv.accounts, argv.async, argv.balance, argv.detached, argv.host);
     }
   )
   .command(
@@ -62,9 +64,10 @@ yargs(hideBin(process.argv))
     (yargs) => {
       CliOptions.addAccountsOption(yargs);
       CliOptions.addBalanceOption(yargs);
+      CliOptions.addAsyncOption(yargs);
     },
     async (argv) => {
-      await HederaUtils.generateAccounts(console, argv.balance, argv.accounts);
+      await HederaUtils.prepareNode(argv.async, console, argv.balance, argv.accounts);
     }
   )
   .command(
@@ -99,9 +102,9 @@ Requirements:
       Disk Image Size: 64 GB`)
   .parse();
 
-async function main(accounts, balance, detached, host) {
+async function main(accounts, async, balance, detached, host) {
   if (detached) {
-    await startDetached(accounts, balance, host);
+    await startDetached(accounts, async, balance, host);
   }
 
   const screen = new TerminalUserInterface();
@@ -111,7 +114,7 @@ async function main(accounts, balance, detached, host) {
   const mirrorNodeLogger = screen.getMirrorNodeLog();
 
   await screen.updateStatusBoard(host);
-  await start(accounts, balance, host, eventLogger, accountLogger);
+  await start(accounts, async, balance, host, eventLogger, accountLogger);
 
   eventLogger.log(
     "\nLocal node has been successfully started. Press Ctrl+C to stop the node."
@@ -176,31 +179,27 @@ function attachContainerLogs(containerId, logger) {
 /**
  * Check if network is up and generate accounts
  */
-async function start(accounts, balance, host, eventLogger, accountLogger) {
+async function start(accounts, async, balance, host, eventLogger, accountLogger) {
   eventLogger.log("Detecting the network...");
   await ConnectionCheck.waitForFiringUp(5600, eventLogger, host);
   await ConnectionCheck.waitForFiringUp(50211, console, host);
   eventLogger.log("Starting the network...");
 
-  eventLogger.log("Importing fees...");
-  await HederaUtils.importFees(host);
-  accountLogger.log("Generating accounts...");
-  await HederaUtils.generateAccounts(accountLogger, balance, accounts, true, host);
+  eventLogger.log("Preparing Node...");
+  await HederaUtils.prepareNode(async, accountLogger, async, balance, accounts, true, host);
 }
 
 /**
  * Check if network is up and generate accounts
  */
-async function startDetached(accounts, balance, host) {
+async function startDetached(accounts, async, balance, host) {
   console.log("Detecting the network...");
   await ConnectionCheck.waitForFiringUp(5600, console, host);
   await ConnectionCheck.waitForFiringUp(50211, console, host);
   console.log("Starting the network...");
 
-  console.log("Importing fees...");
-  await HederaUtils.importFees(host);
-  console.log("Generating accounts...");
-  await HederaUtils.generateAccounts(console, balance, accounts, true, host);
+  console.log("Preparing Node...");
+  await HederaUtils.prepareNode(async, console, balance, accounts, true, host);
   console.log("\nLocal node has been successfully started in detached mode.");
   process.exit();
 }
